+++ projects/angular-sdk/src/i18n/zh.ts
@@ -264,6 +264,7 @@ export const locale: I18n = {
             TERMS_LINK_TEXT: "条款和条件",
             COVERAGE_PERCENTAGE: "承保飞行费用的 {{ coverage }}%",
             PRICE_PER_TRAVELER: "<span>每位乘客</span><span class='price-per-traveler'>{{ price }}</span>",
+            INCL_TAXES: "包括税（如果适用）",
             ERROR_CODE: {
                 CO003: "未找到合同。",
                 CO004: "未找到合同。",