+++ src/app/pages/guides/cfar-offer-dialog/cfar-offer-dialog.component.ts
@@ -46,6 +46,7 @@ export class CfarOfferDialogPageComponent extends CommonGuidesComponent {
         basePath: this.basePath,
         imageBasePath: this.imageBasePath,
         hCSessionId: this.hCSessionId,
+        termsLink: this.termsLink,
         // Choice 1 : The component will create offers based of the itineraries
         itineraries: this.itineraries
         // Choice 2 : The component will load directly the offers
@@ -124,6 +125,13 @@ export class CfarOfferDialogPageComponent extends CommonGuidesComponent {
         required: true
       },
       {
+        name: 'termsLink',
+        description: `
+          Link for Terms and conditions
+        `,
+        required: true
+      },
+      {
         name: 'itineraries',
         description: `
           Itineraries are required in order to create new offers <br />
@@ -172,6 +180,7 @@ export class CfarOfferDialogPageComponent extends CommonGuidesComponent {
       imageBasePath: this.imageBasePath,
       currentLang: this.currentLang,
       hCSessionId: this.hCSessionId,
+      termsLink: this.termsLink,
       itineraries: this.itineraries,
     };
     const dialogConfig = DialogUtils.getDialogConfig(dialogData);